+++ src/components/screens/monitor/delegates/overview/forgingDetails.js
@@ -1,11 +1,8 @@
 // istanbul ignore file
 import React from 'react';
-import { useSelector } from 'react-redux';
-import { Link } from 'react-router-dom';
 import moment from 'moment';
-import { routes, colorPalette, ROUND_LENGTH } from '@constants';
+import { colorPalette, ROUND_LENGTH } from '@constants';
 import { DoughnutChart } from '@toolbox/charts';
-import AccountVisual from '@toolbox/accountVisual';
 import Box from '@toolbox/box';
 import BoxHeader from '@toolbox/box/header';
 import BoxContent from '@toolbox/box/content';
@@ -13,6 +10,7 @@ import BoxEmptyState from '@toolbox/box/emptyState';
 import GuideTooltip, { GuideTooltipItem } from '@toolbox/charts/guideTooltip';
 import Icon from '@toolbox/icon';
 import NumericInfo from './numericInfo';
+import Forger from './forger';
 import styles from './overview.css';
 
 const FORGERS_TO_SHOW = 6;
@@ -30,18 +28,6 @@ const getForgingStats = (data, forgedInRound) => {
   return Object.values(statuses);
 };
 
-const Forger = ({ forger }) => (
-  <div className={`${styles.forger} forger-item`}>
-    <Link to={`${routes.account.path}?address=${forger.address}`}>
-      <AccountVisual
-        address={forger.address}
-        className={styles.accountVisual}
-      />
-      <span>{forger.username}</span>
-    </Link>
-  </div>
-);
-
 const ProgressBar = ({ forgedInRound }) => (
   <div className={styles.progressBar}>
     <div className={styles.lineForged} style={{ width: `${(forgedInRound / ROUND_LENGTH) * 100}%` }} />
@@ -50,24 +36,21 @@ const ProgressBar = ({ forgedInRound }) => (
 
 const formatToTwoDigits = str => str.toLocaleString('en-US', { minimumIntegerDigits: 2 });
 
-const getPassedMinutes = (lastBlock = {}, firstRoundBlock = {}) => {
-  const seconds = lastBlock.timestamp - firstRoundBlock.timestamp;
+const getPassedMinutes = (startTime) => {
+  const seconds = Math.floor((new Date()).getTime() / 1000) - startTime;
   if (!seconds) return '00:00';
   const duration = moment.duration({ seconds });
   return `${formatToTwoDigits(duration.minutes())}:${formatToTwoDigits(duration.seconds())}`;
 };
 
 const ForgingDetails = ({
-  t, forgers,
+  t, forgers, forgedInRound, startTime,
 }) => {
   const delegatesForgedLabels = [
     t('Forging'),
     t('Awaiting slot'),
     t('Missed block'),
   ];
-  const { latestBlocks } = useSelector(state => state.blocks);
-  const forgedInRound = latestBlocks.length
-    ? latestBlocks[0].height % ROUND_LENGTH : 0;
 
   const doughnutChartData = {
     labels: delegatesForgedLabels,
@@ -91,6 +74,7 @@ const ForgingDetails = ({
   };
 
   const forgersListToShow = forgers.slice(0, FORGERS_TO_SHOW);
+  console.log(forgersListToShow);
 
   return (
     <Box className={styles.wrapper}>
@@ -153,7 +137,7 @@ const ForgingDetails = ({
               </section>
               <NumericInfo
                 title="Minutes passed"
-                value={`${getPassedMinutes(latestBlocks[0], latestBlocks[forgedInRound])}`}
+                value={`${getPassedMinutes(startTime)}`}
                 icon="clock"
               />
             </div>
