+++ src/components/screens/monitor/delegates/overview/numericInfo.js
