+++ src/components/screens/monitor/delegates/overview/overview.css
@@ -1,4 +1,4 @@
-@import '../../../../app/mixins.css';
+@import '../../../../../app/mixins.css';
 
 .wrapper {
   & .content {
